// packages/air-ui-ds/models/enums/steppers.ts
export enum StepStatus {
    NONE = 'none',
    INACTIVE = 'inactive',
    CURRENT = 'current',
    COMPLETED = 'completed',
}

export enum StepIndicatorSize {
    XS = 'xs',
    SM = 'sm',
    MD = 'md',
    LG = 'lg',
    XL = 'xl',
}

export enum StepIndicatorType {
    NUMBER = 'number',
    EMPTY = 'empty',
    ICON = 'icon',
}